Use workspace protocol for theme peer dependency
