+++ packages/ag-grid-theme/package.json
@@ -25,6 +25,6 @@
     "del": "^7.0.0"
   },
   "peerDependencies": {
-    "@salt-ds/theme": "^1.23.0"
+    "@salt-ds/theme": "workspace:^"
   }
 }
